split hybrid_count.cpp
