+++ inst/include/dplyr/HybridHandlerMap.h
@@ -8,6 +8,7 @@ typedef dplyr_hash_map<SEXP,HybridHandler> HybridHandlerMap;
 
 void install_simple_handlers(HybridHandlerMap& handlers);
 void install_minmax_handlers(HybridHandlerMap& handlers);
+void install_count_handlers(HybridHandlerMap& handlers);
 
 bool hybridable(RObject arg);
