+++ src/hybrid.cpp
@@ -4,8 +4,6 @@
 
 #include <dplyr/RowwiseDataFrame.h>
 
-#include <dplyr/MultipleVectorVisitors.h>
-
 #include <dplyr/Hybrid.h>
 #include <dplyr/HybridHandlerMap.h>
 
@@ -13,8 +11,6 @@
 #include <dplyr/Result/Rank.h>
 #include <dplyr/Result/ConstantResult.h>
 
-#include <dplyr/Result/Count.h>
-#include <dplyr/Result/Count_Distinct.h>
 #include <dplyr/Result/Lead.h>
 #include <dplyr/Result/Lag.h>
 #include <dplyr/Result/CumSum.h>
@@ -46,42 +42,6 @@ bool hybridable(RObject arg) {
     break;
   }
   return false;
-}
-
-Result* count_prototype(SEXP args, const LazySubsets&, int) {
-  if (Rf_length(args) != 1)
-    stop("n does not take arguments");
-  return new Count;
-}
-
-Result* count_distinct_prototype(SEXP call, const LazySubsets& subsets, int nargs) {
-  MultipleVectorVisitors visitors;
-  bool na_rm = false;
-
-  for (SEXP p = CDR(call); !Rf_isNull(p); p = CDR(p)) {
-    SEXP x = CAR(p);
-    if (!Rf_isNull(TAG(p)) && TAG(p) == Rf_install("na.rm")) {
-      if (TYPEOF(x) == LGLSXP && Rf_length(x) == 1) {
-        na_rm = LOGICAL(x)[0];
-      } else {
-        stop("incompatible value for `na.rm` parameter");
-      }
-    } else if (TYPEOF(x) == SYMSXP) {
-      visitors.push_back(subsets.get_variable(x));
-    } else {
-      return 0;
-    }
-  }
-
-  if (visitors.size() == 0) {
-    stop("need at least one column for n_distinct()");
-  }
-
-  if (na_rm) {
-    return new Count_Distinct_Narm<MultipleVectorVisitors>(visitors);
-  } else {
-    return new Count_Distinct<MultipleVectorVisitors>(visitors);
-  }
 }
 
 Result* row_number_prototype(SEXP call, const LazySubsets& subsets, int nargs) {
@@ -352,8 +312,6 @@ Result* in_prototype(SEXP call, const LazySubsets& subsets, int nargs) {
 HybridHandlerMap& get_handlers() {
   static HybridHandlerMap handlers;
   if (!handlers.size()) {
-    handlers[ Rf_install("n")         ] = count_prototype;
-    handlers[ Rf_install("n_distinct")   ] = count_distinct_prototype;
     handlers[ Rf_install("row_number")   ] = row_number_prototype;
     handlers[ Rf_install("ntile")      ] = ntile_prototype;
 
@@ -379,6 +337,7 @@ HybridHandlerMap& get_handlers() {
 
     install_simple_handlers(handlers);
     install_minmax_handlers(handlers);
+    install_count_handlers(handlers);
   }
   return handlers;
 }
